+++ src/maxtikzlib/figure.py
@@ -8,260 +8,11 @@ import matplotlib.patches as patches
 from maxtikzlib.color import Color
 from maxtikzlib.linestyle import Linestyle
 
-
-class TikzObject:
-    def __init__(
-        self,
-        label: str | None = None,
-        comment: str | None = None,
-        layer: int = 0,
-        **kwargs,
-    ) -> None:
-
-        self._label = label
-        self._comment = comment
-        self._layer = layer
-        self._kwargs = kwargs
-
-    @property
-    def label(self) -> str | None:
-        return self._label
-
-    @property
-    def comment(self) -> str | None:
-        return self._comment
-
-    @property
-    def layer(self) -> int:
-        return self._layer
-
-    @property
-    def kwargs(self) -> dict:
-        return self._kwargs
-
-    @property
-    def options(self) -> str:
-        options = ", ".join(
-            f"{k.replace('_', ' ')}={v}" for k, v in self.kwargs.items()
-        )
-        return options
-
-
-class Tikzlayer:
-    def __init__(self, label, comment=None):
-        self.label = label
-        self.items = []
-
-    def add(self, item):
-        self.items.append(item)
-
-    def get_reqs(self):
-        reqs = set()
-        for item in self.items:
-            if isinstance(item, Path):
-                for node in item.nodes:
-                    if not node.layer == self.label:
-                        reqs.add(node.layer)
-        return reqs
-
-    def generate_tikz(self):
-        tikz_script = f"\n% Layer {self.label}\n"
-        tikz_script += f"\\begin{{pgfonlayer}}{{{self.label}}}\n"
-        for item in self.items:
-            tikz_script += item.to_tikz()
-        tikz_script += f"\\end{{pgfonlayer}}\n"
-        return tikz_script
-
-
-class TikzWrapper:
-    def __init__(self, raw_tikz, label="", content="", layer=0, **kwargs):
-        self.raw_tikz = raw_tikz
-        self.label = label
-        self.content = content
-        self.layer = layer
-        self.options = kwargs
-
-    def to_tikz(self):
-        return self.raw_tikz
-
-
-class Node(TikzObject):
-    def __init__(
-        self,
-        x: float,
-        y: float,
-        label: str = "",
-        content: str = "",
-        comment: str | None = None,
-        layer: int = 0,
-        **kwargs,
-    ):
-        """
-        Represents a TikZ node.
-
-        Parameters:
-        - x (float): X-coordinate of the node.
-        - y (float): Y-coordinate of the node.
-        - name (str, optional): Name of the node. If None, a default name will be assigned.
-        - **kwargs: Additional TikZ node options (e.g., shape, color).
-        """
-        self._x = x
-        self._y = y
-        self._content = content
-        super().__init__(label=label, comment=comment, layer=layer, **kwargs)
-
-    @property
-    def x(self):
-        return self._x
-
-    @property
-    def y(self):
-        return self._y
-
-    @property
-    def content(self):
-        return self._content
-
-    def to_tikz(self):
-        """
-        Generate the TikZ code for this node.
-
-        Returns:
-        - tikz_str (str): TikZ code string for the node.
-        """
-
-        # options = ", ".join(
-        #     f"{k.replace('_', ' ')}={v}" for k, v in self.options.items()
-        # )
-        options = self.options
-        if options:
-            options = f"[{options}]"
-
-        node_string = f"\\node{options} ({self.label}) at ({self.x}, {self.y}) {{{self.content}}};\n"
-        if self.comment is not None:
-            node_string = f"% {self.comment}\n" + node_string
-        return node_string
-
-
-class Path(TikzObject):
-    def __init__(
-        self,
-        nodes,
-        path_actions=[],
-        cycle: bool = False,
-        label: str = "",
-        comment: str | None = None,
-        layer: int = 0,
-        **kwargs,
-    ):
-        """
-        Represents a path (line) connecting multiple nodes.
-
-        Parameters:
-        - nodes (list of str): List of node names to connect.
-        - **kwargs: Additional TikZ path options (e.g., style, color).
-        """
-        self.nodes = nodes
-        self._path_actions = path_actions
-        self._cycle = cycle
-
-        super().__init__(label=label, comment=comment, layer=layer, **kwargs)
-
-    @property
-    def path_actions(self):
-        return self._path_actions
-
-    @property
-    def cycle(self) -> bool:
-        return self._cycle
-
-    def to_tikz(self):
-        """
-        Generate the TikZ code for this path.
-
-        Returns:
-        - tikz_str (str): TikZ code string for the path.
-        """
-        # options = ", ".join(
-        #     f"{k.replace('_', ' ')}={v}" for k, v in self.options.items()
-        # )
-
-        options = self.options
-        if len(self.path_actions) > 0:
-            options = ", ".join(self.path_actions) + ", " + options
-        if options:
-            options = f"[{options}]"
-        path_str = " to ".join(f"({node.label}.center)" for node in self.nodes)
-        if self.cycle:
-            path_str += " -- cycle"
-
-        path_str = f"\\draw{options} {path_str};\n"
-
-        if self.comment is not None:
-            path_str = f"% {self.comment}\n" + path_str
-
-        return path_str
-
-
-class Loop(TikzObject):
-    def __init__(self, variable, values, layer=0):
-        self._variable = variable
-        self._values = list(values)
-        self._items = []
-
-        super().__init__(layer=layer)
-
-    @property
-    def variable(self):
-        return self._variable
-
-    @property
-    def values(self):
-        return self._values
-
-    @property
-    def items(self):
-        return self._items
-
-    def add_node(self, *args, **kwargs):
-        node = Node(*args, **kwargs)
-        self._items.append(node)
-        return node
-
-    def add_path(self, nodes, comment: str | None = None, **kwargs):
-        path = Path(nodes, comment=comment, **kwargs)
-        self._items.append(path)
-        return path
-
-    def add_raw(self, raw_tikz):
-        wrapper = TikzWrapper(raw_tikz)
-        self._items.append(wrapper)
-        return wrapper
-
-    def to_tikz(self):
-        values_str = ",".join(str(v) for v in self.values)
-        tikz_body = "".join([item.to_tikz() for item in self.items])
-        return f"\\foreach \\{self.variable} in {{{values_str}}}{{% start \\foreach\n{tikz_body}}}% end \\foreach\n"
-
-
-class TikzLoopContext:
-    def __init__(self, figure, variable, values, layer=0):
-        self._figure = figure
-        self._loop = Loop(variable, values, layer=layer)
-
-    @property
-    def figure(self):
-        return self._figure
-
-    @property
-    def loop(self):
-        return self._loop
-
-    def __enter__(self):
-        return self.loop
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        self.figure.add_item(self.loop, self.loop.layer)
+from maxtikzlib.node import Node
+from maxtikzlib.path import Path
+from maxtikzlib.wrapper import TikzWrapper
+from maxtikzlib.layer import Tikzlayer
+from maxtikzlib.loop import TikzLoopContext
 
 
 class TikzFigure:
