+++ src/maxtikzlib/node.py
@@ -1,0 +1,59 @@
+from maxtikzlib.base import TikzObject
+
+
+class Node(TikzObject):
+    def __init__(
+        self,
+        x: float,
+        y: float,
+        label: str = "",
+        content: str = "",
+        comment: str | None = None,
+        layer: int = 0,
+        **kwargs,
+    ):
+        """
+        Represents a TikZ node.
+
+        Parameters:
+        - x (float): X-coordinate of the node.
+        - y (float): Y-coordinate of the node.
+        - name (str, optional): Name of the node. If None, a default name will be assigned.
+        - **kwargs: Additional TikZ node options (e.g., shape, color).
+        """
+        self._x = x
+        self._y = y
+        self._content = content
+        super().__init__(label=label, comment=comment, layer=layer, **kwargs)
+
+    @property
+    def x(self):
+        return self._x
+
+    @property
+    def y(self):
+        return self._y
+
+    @property
+    def content(self):
+        return self._content
+
+    def to_tikz(self):
+        """
+        Generate the TikZ code for this node.
+
+        Returns:
+        - tikz_str (str): TikZ code string for the node.
+        """
+
+        # options = ", ".join(
+        #     f"{k.replace('_', ' ')}={v}" for k, v in self.options.items()
+        # )
+        options = self.options
+        if options:
+            options = f"[{options}]"
+
+        node_string = f"\\node{options} ({self.label}) at ({self.x}, {self.y}) {{{self.content}}};\n"
+        if self.comment is not None:
+            node_string = f"% {self.comment}\n" + node_string
+        return node_string
